bot/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
motor==3.3.1
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pydantic==1.10.13
pymongo==4.5.0
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.8.0
yarl==1.9.2
